pyproject.toml
[project]
name = "recom-azure-function"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "azure-functions-durable>=1.2.10",
    "azure-functions>=1.22.1",
    "azure-storage-blob>=12.25.1",
    "beautifulsoup4>=4.13.3",
    "openai>=1.71.0",
    "pandas>=2.2.3",
    "playwright>=1.51.0",
    "pydantic>=2.11.2",
    "requests>=2.32.3",
]
